pyproject.toml
[tool.poetry]
name = "home_work10"
version = "0.1.0"
description = ""
authors = ["SaifuliAnna <saifulianna.it@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.2.0"
mongoengine = "^0.24.2"
redis-lru = "^0.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
